completed CCC 2012 S5, back to school 19 P3

+++ DMOJ Contests (Live)/Back To School 19/P3.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 int main() {
-    double n, m, k;
+    long double n, m, k;
     cin >> n >> m;
     k = ceil(ceil(log2(n))/m);
     cout << k;
